feat: add PluginsLib and fix cyrillic letters
