+++ practice2025/practice2025.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35506.116 d17.12
+VisualStudioVersion = 17.12.35506.116
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "task10", "task10\task10.csproj", "{A6BA607E-A894-4C88-B062-0685D2B8DAD8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "task10tests", "task10tests\task10tests.csproj", "{ACD7CD7D-A4A7-4E0A-8648-0A1A14590B31}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "plugin", "plugin\plugin.csproj", "{56EA84E9-FE0F-43EA-8A68-78FC33CC1D71}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PluginsLib", "PluginsLib\PluginsLib.csproj", "{5FB4447F-D67C-4005-8B20-4C75D55FDB3A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{56EA84E9-FE0F-43EA-8A68-78FC33CC1D71}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{56EA84E9-FE0F-43EA-8A68-78FC33CC1D71}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{56EA84E9-FE0F-43EA-8A68-78FC33CC1D71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5FB4447F-D67C-4005-8B20-4C75D55FDB3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5FB4447F-D67C-4005-8B20-4C75D55FDB3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5FB4447F-D67C-4005-8B20-4C75D55FDB3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5FB4447F-D67C-4005-8B20-4C75D55FDB3A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
